LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/redirect/same-origin-redirect.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/record-header.py?file=redirect-same-origin-same-site2f392904-96f5-4a10-862e-11d7b033d534


Harness Error (TIMEOUT), message = null

PASS Same-Origin -> Same-Origin redirect
TIMEOUT Same-Origin -> Same-Site redirect Test timed out
NOTRUN Same-Origin -> Cross-Site redirect
FAIL undefined: sec-fetch-dest assert_equals: expected "image" but got ""
FAIL undefined: sec-fetch-mode assert_equals: expected "no-cors" but got ""
FAIL undefined: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS undefined: sec-fetch-user
